Complete plan items: security fixes, CLI access, WIP items
Security fixes:
- Fix analyzer.rs compilation error (loop variable mutation)
- Add SensitiveString support to proxy credentials (SocksProxy, HttpConnectProxy)
- Add SensitiveString.len(), as_bytes(), is_empty() methods
- Add template signature verification to marketplace downloads

CLI Access:
- Implement config validate/show commands
- Wire up vuln module CLI access (CVSS scoring, exploitability, prioritization)
- Wire up storage module CLI access
- Fix notify send command integration

Code Quality:
- Gate unused import in distributed/io.rs on #[cfg(test)]
- Add proper impl methods to vuln module types

Feature Improvements:
- Add TLS config options to ServeArgs
- Feature-gate AgentRegistry routes on ai-integration
- Add MCP constraint context for LLM tool calls

Vuln Module Enhancements:
- CvssScore: base_score(), severity(), temporal_score()
- ExploitInfo: for_cve(), exploitability_score(), has_public_exploit()
- RiskScore: new(), total(), priority()
- TriageResult: new(), status()
- Remediation: from_severity(), priority(), effort(), steps()

+++ crates/slapper/src/agent/portfolio.rs
@@ -9,7 +9,7 @@ use std::path::PathBuf;
 use std::sync::{Arc, RwLock};
 
 use anyhow::Result;
-use chrono::{DateTime, Utc};
+use chrono::{DateTime, Utc, Timelike};
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
@@ -77,14 +77,14 @@ pub struct OffPeakWindow {
 impl OffPeakWindow {
     pub fn is_in_window(&self, time: &DateTime<Utc>) -> bool {
         let local = match &self.timezone[..] {
-            "UTC" => time.hour(),
+            "UTC" => time.hour() as i32,
             _ => {
-                let offset_hours: i32 = self.timezone.trim().parse().unwrap_or(0);
+                let offset_hours: i64 = self.timezone.trim().parse().unwrap_or(0);
                 let offset_time = *time + chrono::Duration::hours(offset_hours);
                 offset_time.hour() as i32
             }
         };
-        let current_hour = local as i32;
+        let current_hour = local;
         let start = self.start_hour as i32;
         let end = self.end_hour as i32;
